+++ py-rattler/src/error.rs
@@ -4,8 +4,8 @@ use pyo3::{create_exception, exceptions::PyException, PyErr};
 use rattler::install::TransactionError;
 use rattler_conda_types::{
     ConvertSubdirError, InvalidPackageNameError, ParseArchError, ParseChannelError,
-    ParseMatchSpecError, ParsePlatformError, ParseVersionError, VersionBumpError,
-    VersionExtendError,
+    ParseMatchSpecError, ParsePlatformError, ParseVersionError, ValidatePackageRecordsError,
+    VersionBumpError, VersionExtendError,
 };
 use rattler_lock::{ConversionError, ParseCondaLockError};
 use rattler_package_streaming::ExtractError;
@@ -70,6 +70,12 @@ pub enum PyRattlerError {
     GatewayError(#[from] GatewayError),
     #[error(transparent)]
     InstallerError(#[from] rattler::install::InstallerError),
+    #[error(transparent)]
+    ParseExplicitEnvironmentSpecError(
+        #[from] rattler_conda_types::ParseExplicitEnvironmentSpecError,
+    ),
+    #[error(transparent)]
+    ValidatePackageRecordsError(#[from] ValidatePackageRecordsError),
 }
 
 fn pretty_print_error(mut err: &dyn Error) -> String {
@@ -154,6 +160,12 @@ impl From<PyRattlerError> for PyErr {
             PyRattlerError::InstallerError(err) => {
                 InstallerException::new_err(pretty_print_error(&err))
             }
+            PyRattlerError::ParseExplicitEnvironmentSpecError(err) => {
+                ParseExplicitEnvironmentSpecException::new_err(pretty_print_error(&err))
+            }
+            PyRattlerError::ValidatePackageRecordsError(err) => {
+                ValidatePackageRecordsException::new_err(pretty_print_error(&err))
+            }
         }
     }
 }
@@ -184,3 +196,9 @@ create_exception!(exceptions, ExtractException, PyException);
 create_exception!(exceptions, ActivationScriptFormatException, PyException);
 create_exception!(exceptions, GatewayException, PyException);
 create_exception!(exceptions, InstallerException, PyException);
+create_exception!(
+    exceptions,
+    ParseExplicitEnvironmentSpecException,
+    PyException
+);
+create_exception!(exceptions, ValidatePackageRecordsException, PyException);