+++ crates/rattler_package_streaming/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.22.11](https://github.com/conda/rattler/compare/rattler_package_streaming-v0.22.10...rattler_package_streaming-v0.22.11) - 2024-11-04
+
+### Other
+
+- release ([#903](https://github.com/conda/rattler/pull/903))
+
 ## [0.22.10](https://github.com/conda/rattler/compare/rattler_package_streaming-v0.22.9...rattler_package_streaming-v0.22.10) - 2024-10-07
 
 ### Added